refactor: Change structure
