+++ apps/easy-ui/src/index.ts
@@ -1,9 +1,9 @@
 export * from '@mui/material';
 
-export * from './EzForm';
+export * from './form';
 
-export * from './EzButton';
-export * from './EzCheckbox';
-export * from './EzIcon';
-export * from './EzTextField';
+export * from './button';
+export * from './checkbox';
+export * from './icon';
+export * from './textfield';
 export * from './toast';
